🐛 fix: invisible color bug
sometimes, the color of the image would be approximated to the same color as the invisible color. even though this approximation is correct, it was treated as a bug because RGB(6, 7, 7) is the code of an invisible color and is not shown on screen. this bug is now fixed, and the color is rounded to RGBA(7, 7, 7, 1) (white).